Fixed layout and structure issues in user dashboard components
