+++ src/App.css
@@ -1,24 +1,25 @@
-/* App.css */
 .app-container {
   display: flex;
-  height: 100vh;
+  width: 100vw; /* spanning the full viewport width */
+  margin: 0;
+  padding: 0;
 }
-
 
 .sidebar {
   width: 20%;
-  height: 105%;
-  background-color: rgb(235, 235, 235);
+  background-color: white;
   color: rgb(55, 53, 47);
   padding: 20px;
   display: flex;
   flex-direction: column;
-  transition: width 0.3s ease;
+  transition: width 0.5s ease;
 }
 
 .sidebar.closed {
   width: 60px;
+  justify-content: center;
 }
+
 
 .sidebar-header {
   display: flex;
@@ -28,7 +29,8 @@
 }
 
 .sidebar-header h2 {
-  font-size: 20px;
+  font-size: 1.5em;
+  font-weight: bold;
   margin: 0;
   white-space: nowrap;
   overflow: hidden;
@@ -44,7 +46,7 @@
 
 .sidebar ul li {
   margin-bottom: 15px;
-  background-color: rgb(235, 235, 235);
+  background-color: white;
   padding: 10px;
   border-radius: 5px;
   transition: background-color 0.3s ease, padding-left 0.3s ease;
@@ -60,14 +62,14 @@
 }
 
 .sidebar ul li:hover {
-  background-color: rgb(225, 225, 225);
+  background-color:	#f1f1f1;
   color: black;
   padding-left: 15px;
 }
 
 .sidebar ul li:active,
 .sidebar ul li a.active {
-  background-color: rgb(175, 175, 175);
+  background-color:	#d6d6d6;
 }
 
 .sidebar ul li:active .sidebar-icon,
@@ -107,14 +109,18 @@
 }
 
 
-/* Main Content Styling */
+/* Main Content */
 .main-content {
-  width: 100%;
-  height: 105%;
+  width: 100%; 
   padding: 20px;
   background-color: rgb(255, 255, 255);
   color: rgb(55, 53, 47);
+  overflow-y: auto;
+  flex-grow: 1;
+  overflow: auto;
+  padding: 20px;
 }
+
 
 .header {
   margin-bottom: 20px;
@@ -122,12 +128,12 @@
 }
 
 .header h1 {
-  font-size: 28px;
+  font-size: 2em;
   font-weight: bold;
 }
 
 .header p {
-  font-size: 16px;
+  font-size: 1.2em;
   font-weight: bold;
   color: rgb(55, 53, 47);
 }
@@ -139,14 +145,15 @@
 }
 
 .card {
-  background-color: rgb(235, 235, 235);
-  padding: 20px;
+  background-color: #f9f9f9;
+  padding: 10px;
+  margin: 10px;
   border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.1);
   display: flex;
   flex-direction: column; /* Stack child elements vertically */
   position: relative;
-  padding-bottom: 30px; /* Add some space at the bottom for the link */
+  padding-bottom: 30px;
 }
 
 .card h3 {
@@ -170,121 +177,32 @@
   margin-bottom: 10px;
 }
 
-.card button {
-  background-color: rgb(0, 0, 0);
-  color: rgb(255, 255, 255);
-  padding: 10px 15px;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-  align-self: center; /* Center the button horizontally within the flex container */
-  margin-top: auto; /* Push the button to the bottom of the card */
+:root {
+  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+  /* line-height: 1.3; */
+  font-weight: 400;
+
+  font-synthesis: none;
+  text-rendering: optimizeLegibility;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
 }
 
-.card button:hover {
-  background-color: rgb(175, 175, 175);
-  color: rgb(55, 53, 47);
-}
-
-.card button:active {
-  background-color: rgb(255, 255, 255);
-  color: rgb(0, 0, 0);
-}
-
-/* Expandable card */
-
-.read-more-link {
-  color: #007bff;
-  cursor: pointer;
-  font-size: 14px;
-  text-decoration: none;
-  transition: color 0.3s ease;
-}
-
-.read-more-link:hover {
-  color: #0056b3;
-  text-decoration: underline;
-}
-
-.card .read-more-link {
-  position: absolute;
-  bottom: 10px;
-  left: 20px;
+a {
+  font-weight: bold;
+  text-decoration: inherit;
 }
 
 
-/* Navbar Styles */
-.navbar {
-  display: none; /* Hide by default */
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  background-color: rgb(235, 235, 235);
-  color: rgb(55, 53, 47);
-  padding: 20px;
-  z-index: 1000;
-}
-
-.navbar-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.navbar-header h2 {
-  margin: 0;
-  font-size: 18px;
-}
-
-.navbar-menu {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-  max-height: 0;
-  overflow: hidden;
-  transition: max-height 0.3s ease;
-}
-
-.navbar-menu.open {
-  max-height: 5%;
-}
-
-.navbar-menu li {
-  margin: 18px 0;
-}
-
-.navbar-menu li a {
-  display: flex;
-  align-items: center;
-  text-decoration: none;
-  color: rgb(55, 53, 47);
-}
-
-.navbar-icon {
-  margin-right: 10px;
-}
-
-/* Responsive Layout */
-@media (max-width: 768px) {
-  .navbar {
-    display: block;
+@media (prefers-color-scheme: light) {
+  :root {
+    color: #213547;
+    background-color: #ffffff;
   }
-
-  .sidebar {
-    display: none;
+  a:hover {
+    color: #747bff;
   }
-
-  .main-content {
-    margin-top: 60px; /* Adjust based on navbar height */
-  }
-
-  .app-container {
-    flex-direction: column;
-  }
-
-  .dashboard-content {
-    grid-template-columns: 1fr;
+  button {
+    background-color: #f9f9f9;
   }
 }
