+++ src/views/components/IncidentAlert.css
@@ -1,81 +1,144 @@
-body {
-    font-family: Arial, sans-serif;
-    background-color: #f4f4f4;
-    margin: 0;
-    padding: 20px;
+.incident-alert-container {
     display: flex;
-    justify-content: center;
+    flex-direction: column;
     align-items: center;
-    height: 100vh;
-    background-image: url('/public/images/wits\ picture.jpg');
-    background-size: cover; /* Cover the entire page */
-    background-position: center; 
-    background-attachment: fixed; /* Keeps the image fixed when scrolling */
-    background-repeat: no-repeat;
-}
-.container {
-    background-color: #fff;
+    justify-content: center;
+    min-height: calc(100vh - 40px); /* Adjust based on your actual header/footer height */
     padding: 20px;
-    border-radius: 20px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    max-width: 400px;
+    background-color: #f9f9f9;
+  }
+  
+  .incident-form {
+    background-color: white;
+    border-radius: 12px;
+    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.1);
+    padding: 40px;
     width: 100%;
-    border: 2px solid black;
-}
-h1 {
-    text-align: center;
+    max-width: 600px;
+    transition: all 0.3s ease;
+  }
+  
+  .incident-form:hover {
+    transform: translateY(-5px);
+    box-shadow: 0 6px 25px rgba(0, 0, 0, 0.15);
+  }
+  
+  h1 {
     color: #333;
-}
-h2 {
-    text-align: center;
-    color: red;
-}
-.input-group {
-    margin-bottom: 15px;
-}
-.input-group label {
-    display: block;
-    margin-bottom: 5px;
-    color: #555;
-}
-.input-group input,
-.input-group textarea {
-    width: 100%; /* Ensure the textarea takes up the full width of its container */
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 4px; 
-    height: 100px;
-    box-sizing: border-box; /* Includes padding and border in the element's total width and height */
-    padding: 10px;
-    resize: vertical; /* gives priority to vertical sizing rather than horizontal sizing */
-}
-.input-group textarea {
-    resize: vertical;
-    height: 100px;
-}
-.input-group input[type="file"] {
-    padding: 3px;
-}
-.buttons {
-    text-align: center;
-}
-.buttons button {
-    background-color: #007BFF;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    margin: 5px;
-}
-.buttons button:hover {
-    background-color: #0056b3;
-}
-#incidentType, #incidentDescription {
-    width: 100%;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    box-sizing: border-box;
+    font-size: 28px;
     margin-bottom: 10px;
-}
+    text-align: center;
+  }
+  
+  h2 {
+    color: #e74c3c;
+    font-size: 18px;
+    margin-bottom: 30px;
+    text-align: center;
+  }
+  
+  .input-group {
+    margin-bottom: 25px;
+    position: relative;
+  }
+  
+  .input-group label {
+    display: block;
+    margin-bottom: 8px;
+    color: #555;
+    font-size: 14px;
+    font-weight: 600;
+  }
+  
+  .input-group select,
+  .input-group textarea,
+  .input-group input[type="file"] {
+    width: 100%;
+    padding: 12px;
+    border: 2px solid #ddd;
+    border-radius: 8px;
+    font-size: 16px;
+    transition: all 0.3s ease;
+    background-color: #f8f8f8;
+  }
+  
+  .input-group select:focus,
+  .input-group textarea:focus {
+    border-color: #4a90e2;
+    outline: none;
+    background-color: #fff;
+  }
+  
+  .input-group textarea {
+    resize: vertical;
+    min-height: 120px;
+  }
+  
+  .input-group input[type="file"] {
+    padding: 10px;
+    background-color: #f0f0f0;
+    cursor: pointer;
+  }
+  
+  .buttons {
+    display: flex;
+    justify-content: space-between;
+    margin-top: 30px;
+  }
+  
+  .buttons button {
+    flex: 1;
+    margin: 0 10px;
+    padding: 12px 20px;
+    border: none;
+    border-radius: 8px;
+    font-size: 16px;
+    font-weight: 600;
+    cursor: pointer;
+    transition: all 0.3s ease;
+    text-transform: uppercase;
+  }
+  
+  .buttons button[type="submit"] {
+    background-color: #4a90e2;
+    color: white;
+  }
+  
+  .buttons button[type="submit"]:hover {
+    background-color: #357ae8;
+  }
+  
+  .buttons button[type="button"] {
+    background-color: #e74c3c;
+    color: white;
+  }
+  
+  .buttons button[type="button"]:hover {
+    background-color: #c0392b;
+  }
+  
+  .buttons button:disabled {
+    background-color: #cccccc;
+    cursor: not-allowed;
+  }
+  
+  .loading-indicator {
+    text-align: center;
+    margin-top: 20px;
+    font-size: 16px;
+    color: #4a90e2;
+  }
+  
+  @media (max-width: 768px) {
+    .incident-form {
+      padding: 30px;
+    }
+    
+    .buttons {
+      flex-direction: column;
+    }
+    
+    .buttons button {
+      margin: 10px 0;
+    }
+  }
