+++ src/views/pages/SignUpPage.css
@@ -2,7 +2,7 @@
 body {
   margin: 0;
   font-family: 'Arial', sans-serif;
-  background-color: #e6f2ff;
+  background-color: white;
   display: flex;
   justify-content: center;
   align-items: center;
