+++ src/views/pages/UserDashboard.css
@@ -183,7 +183,7 @@
 
 .card button:hover {
   background-color: rgb(175, 175, 175);
-  color: rgb(55, 53, 47);
+  color: white;
 }
 
 .card button:active {
@@ -207,7 +207,6 @@
 }
 
 .card .read-more-link {
-  position: absolute;
   bottom: 10px;
   left: 20px;
 }
@@ -286,4 +285,129 @@
   .dashboard-content {
     grid-template-columns: 1fr;
   }
+}
+
+
+/* Incident Reporting Button */
+.incident-reporting h3{
+  font-size: 1.5em;
+  font-weight: bold;
+  text-align: center;
+
+}
+
+.incident-reporting button {
+  width: 70%;
+  padding: 10px;
+  margin-top: 25px;
+  background-color: #4a90e2;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  align-self: center;
+}
+
+.incident-reporting button:hover{
+  background-color: #357ae8;
+}
+
+.incident-reporting button:active{
+  background-color: #3c5c87;
+}
+
+.incident-reporting form {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.incident-reporting textarea {
+  width: 100%;
+  height: 100px;
+  resize: vertical;
+  background-color: #d0d0d0;
+  color: #333;
+  border-radius: 4px;
+  border: none;
+}
+
+.incident-reporting p{
+  font-size: 14px;
+  text-align: center;
+  margin-top: 28px;
+}
+
+
+
+/* Safety Resources Section */
+.safety-resources {
+  height: auto;
+  text-align: center;
+  background-color: #f9f9f9;
+  padding: 10px;
+  margin: 10px;
+  border-radius: 8px;
+  box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.1);
+}
+.safety-resources h3{
+    font-size: 1.5em;
+    font-weight: bold;
+    text-align: center;
+}
+.safety-resources h4{
+    font-size: 1.1em;
+    font-weight: bold;
+    text-align: center;
+}
+
+.safety-resources ul {
+  text-align:start;
+  padding-left: 20px;
+  margin-bottom: 15px;
+}
+
+.safety-resources button {
+  width: 70%;
+  padding: 10px;
+  margin-top: 10px;
+  margin-bottom: 10px;
+  background-color: #4a90e2;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  align-self: center;
+}
+
+.emergency-contacts button {
+  width: 40%;
+  padding: 10px;
+  margin-top: 10px;
+  margin-bottom: 10px;
+  background-color: #4a90e2;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  align-self: center;
+}
+
+.safety-tips {
+  height: auto;
+  text-align: center;
+  background-color: #f9f9f9;
+  padding: 10px;
+  margin: 10px;
+  border-radius: 8px;
+  box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.1);
+  align-items: center;
+}
+.safety-resources button:hover{
+  background-color: 	#357ae8;
+}
+
+.safety-resources button:active{
+  background-color: 	#3c5c87;
+
 }
